feat: fix login route not found, adopt Laravel >= 11 structure
