+++ bootstrap/app.php
@@ -1,55 +1,56 @@
 <?php
 
-/*
-|--------------------------------------------------------------------------
-| Create The Application
-|--------------------------------------------------------------------------
-|
-| The first thing we will do is create a new Laravel application instance
-| which serves as the "glue" for all the components of Laravel, and is
-| the IoC container for the system binding all of the various parts.
-|
-*/
+use Illuminate\Foundation\Application;
+use Illuminate\Foundation\Configuration\Exceptions;
+use Illuminate\Foundation\Configuration\Middleware;
 
-$app = new Illuminate\Foundation\Application(
-    $_ENV['APP_BASE_PATH'] ?? dirname(__DIR__)
-);
+return Application::configure(basePath: dirname(__DIR__))
+    ->withProviders()
+    ->withRouting(
+        web: __DIR__.'/../routes/web.php',
+        commands: __DIR__.'/../routes/console.php',
+        health: '/up',
+    )
+    ->withMiddleware(function (Middleware $middleware) {
+        $middleware->encryptCookies(except: [
+            'theme',
+        ]);
 
-/*
-|--------------------------------------------------------------------------
-| Bind Important Interfaces
-|--------------------------------------------------------------------------
-|
-| Next, we need to bind some important interfaces into the container so
-| we will be able to resolve them when needed. The kernels serve the
-| incoming requests to this application from both the web and CLI.
-|
-*/
+        $middleware->preventRequestsDuringMaintenance(except: [
+            '/status/db',
+            '/status',
+        ]);
 
-$app->singleton(
-    Illuminate\Contracts\Http\Kernel::class,
-    App\Http\Kernel::class
-);
+        $middleware->redirectGuestsTo(fn () => localized_route('login'));
 
-$app->singleton(
-    Illuminate\Contracts\Console\Kernel::class,
-    App\Console\Kernel::class
-);
+        $middleware->throttleApi();
 
-$app->singleton(
-    Illuminate\Contracts\Debug\ExceptionHandler::class,
-    App\Exceptions\Handler::class
-);
+        $middleware->trimStrings(except: [
+            'password',
+            'password_confirmation',
+        ]);
 
-/*
-|--------------------------------------------------------------------------
-| Return The Application
-|--------------------------------------------------------------------------
-|
-| This script returns the application instance. The instance is given to
-| the calling script so we can separate the building of the instances
-| from the actual running of the application and sending responses.
-|
-*/
+        $middleware->trustHosts(at: fn () => config('app.url'), subdomains: true);
 
-return $app;
+        $middleware->trustProxies(at: '*');
+
+        $middleware->prependToGroup('web', [
+            \ChinLeung\MultilingualRoutes\DetectRequestLocale::class,
+        ]);
+
+        $middleware->appendToGroup('web', [
+            \RalphJSmit\Livewire\Urls\Middleware\LivewireUrlsMiddleware::class,
+            \App\Http\Middleware\ConfirmLanguage::class,
+        ]);
+
+        $middleware->alias([
+            'admin' => \App\Http\Middleware\AdminMiddleware::class,
+            'password.confirm' => \App\Http\Middleware\RequirePassword::class,
+            'verified' => \App\Http\Middleware\EnsureEmailIsVerified::class,
+            'localize' => \App\Http\Middleware\RedirectToPreferredLocale::class,
+            'onboard' => \App\Http\Middleware\RedirectForOnboarding::class,
+        ]);
+    })
+    ->withExceptions(function (Exceptions $exceptions) {
+        //
+    })->create();
