+++ bootstrap/providers.php
@@ -1,0 +1,7 @@
+<?php
+
+return [
+    App\Providers\AppServiceProvider::class,
+    App\Providers\FortifyServiceProvider::class,
+    App\Providers\Filament\AdminPanelProvider::class,
+];
