+++ config/filesystems.php
@@ -1,52 +1,11 @@
 <?php
 
 return [
-
-    /*
-    |--------------------------------------------------------------------------
-    | Default Filesystem Disk
-    |--------------------------------------------------------------------------
-    |
-    | Here you may specify the default filesystem disk that should be used
-    | by the framework. The "local" disk, as well as a variety of cloud
-    | based disks are available to your application. Just store away!
-    |
-    */
-
-    'default' => env('FILESYSTEM_DRIVER', 'local'),
-
-    /*
-    |--------------------------------------------------------------------------
-    | Filesystem Disks
-    |--------------------------------------------------------------------------
-    |
-    | Here you may configure as many filesystem "disks" as you wish, and you
-    | may even configure multiple disks of the same driver. Defaults have
-    | been setup for each driver as an example of the required options.
-    |
-    | Supported Drivers: "local", "ftp", "sftp", "s3"
-    |
-    */
-
     'disks' => [
-
-        'local' => [
-            'driver' => 'local',
-            'root' => storage_path('app'),
-        ],
-
-        'public' => [
-            'driver' => 'local',
-            'root' => storage_path('app/public'),
-            'url' => '/storage',
-            'visibility' => 'public',
-        ],
-
         'snapshots' => [
             'driver' => 'local',
             'root' => database_path('snapshots'),
         ],
-
         'snapshots-s3' => [
             'driver' => 's3',
             'key' => env('SNAPSHOTS_AWS_ACCESS_KEY_ID'),
@@ -56,7 +15,6 @@ return [
             'url' => env('SNAPSHOTS_AWS_URL'),
             'endpoint' => env('SNAPSHOTS_AWS_ENDPOINT'),
         ],
-
         's3' => [
             'driver' => 's3',
             'key' => env('AWS_ACCESS_KEY_ID'),
@@ -66,27 +24,12 @@ return [
             'url' => env('AWS_URL'),
             'endpoint' => env('AWS_ENDPOINT'),
         ],
-
     ],
-
-    /*
-    |--------------------------------------------------------------------------
-    | Symbolic Links
-    |--------------------------------------------------------------------------
-    |
-    | Here you may configure the symbolic links that will be created when the
-    | `storage:link` Artisan command is executed. The array keys should be
-    | the locations of the links and the values should be their targets.
-    |
-    */
-
     'links' => [
-        public_path('storage') => storage_path('app/public'),
         public_path('sitemap.xml') => storage_path('app/public/sitemap.xml'),
         public_path('robots.txt') => storage_path('app/public/robots.txt'),
         lang_path('lsq') => lang_path('fr'),
         lang_path('lsq.json') => lang_path('fr.json'),
         lang_path('vendor/hearth-components/lsq') => lang_path('vendor/hearth-components/fr'),
     ],
-
 ];
