+++ tests/Feature/View/Components/InterpretationTest.php
@@ -169,10 +169,10 @@ test('do not fallback to LSQ (lsq)', function () {
 });
 
 test('no Interpretation without sign language translations setting enabled', function () {
-get(localized_route('welcome'))
-->assertStatus(200);
+    get(localized_route('welcome'))
+        ->assertStatus(200);
 
     $interpretations = Interpretation::all();
 
     expect($interpretations)->toHaveCount(0);
-    });
+});
